% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/br_pol.R
\name{br_pol}
\alias{br_pol}
\title{Get Sex of Person Based on First Name Using Brazilian Politician Data}
\usage{
br_pol(names = NULL, matchAccent = TRUE)
}
\arguments{
\item{names}{vector of names}

\item{matchAccent}{boolean for whether to match accent. Not matching on accent may be slower.}
}
\value{
data.frame with original list and proportion women with that name
}
\description{
Fetches proportion of women with that first name in brazilian politician data.
}
\examples{
br_pol(names="Antonio")

}

