% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/br_pol_model.R
\name{br_pol_model}
\alias{br_pol_model}
\title{Get Sex of Person Based on Model Trained on Brazilian Politicians}
\usage{
br_pol_model(names = NULL)
}
\arguments{
\item{names}{vector of names}
}
\value{
data.frame with original list and proportion women with that name
}
\description{
Get predicted sex of a Portuguese name based on a model trained on a dataset
of brazilian politicians. This model uses character counts, first and last
characters, and bigrams. Note the predictions favor men as most of the
database of politician's are men. With weak prediction the model favors
just guessing male.
}
\examples{
br_pol_model(names="Antonio")

}

