% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{clean_names}
\alias{clean_names}
\title{Remove Whitespace, Accents, Uppercase}
\usage{
clean_names(names = NULL, encoding = "ASCII//TRANSLIT")
}
\arguments{
\item{names}{vector of names}
}
\value{
vector of cleaned names
}
\description{
Remove Whitespace, Accents, Uppercase
}

