% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{gen_features}
\alias{gen_features}
\title{Generate Features for ML model of classification}
\usage{
gen_features(names = NULL)
}
\arguments{
\item{names}{vector of names}
}
\value{
matrix that is used in our basic ML classification model
}
\description{
Generates features and appropriate data matrix to be used in prediction using existing models. Ignores accents for now.
}
\examples{
br_pol(names="dilma")

}

