% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{ngram}
\alias{ngram}
\title{Generates ngrams of characters}
\usage{
ngram(xs, n = 2)
}
\arguments{
\item{n}{'n'-grams to generate}

\item{string}{string to be chunked}
}
\value{
character vector of all ngrans
}
\description{
Generates ngrams of characters
}

