% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{gen_features}
\alias{gen_features}
\title{Generate Features for ML model of classification}
\usage{
gen_features(names = NULL)
}
\arguments{
\item{names}{vector of names}
}
\value{
matrix that is used in our basic ML classification model. May be have more columns than the fit model. Make sure to only keep columns found in the model.
}
\description{
Do not use this unless you know exactly what it does. Use the prediction functions instead. Generates features and appropriate data matrix to be used in prediction using existing models. Ignores accents for now.
}

